CMT 002.1 - Fix @id @generatedvalue gap

+++ ois/src/main/java/com/tim3/ois/repository/UserRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
-    User findById(Long id);
+    User findById(int id);
 
 }
